-getSize now returns an `unsigned long long`.
This type now matches the type and width of the value returned by the
`fileSize` accessor of the `NSDictionary (NSFileAttributes)` category
(which is what is used when accumulating the total cache size).

+++ SDWebImage/SDImageCache.h
@@ -134,7 +134,7 @@ typedef enum SDImageCacheType SDImageCacheType;
 /**
  * Get the size used by the disk cache
  */
-- (int)getSize;
+- (unsigned long long)getSize;
 
 /**
  * Get the number of images in the disk cache
